+++ Spellchecking/src/spell_checker.py
@@ -1,3 +1,9 @@
+""" Spellchecker is the code doing the checking of 
+words whether they follow the english language 
+"""
+import dataclasses
+
+@dataclasses.dataclass
 class TrieNode:
     """A node in the trie structure"""
 
@@ -16,25 +22,27 @@ class TrieNode:
         # keys are characters, values are nodes
         self.children = {}
 
-class SpellChecker(object):
-    """Spell checking interface - instanisiated with sc = Spellchecker(wordList: string)"""
+@dataclasses.dataclass
+class SpellChecker:
+    """Spell checking interface - instanisiated with sc = Spellchecker(word_list: string)"""
 
-    def __init__(self, wordList):
+    def __init__(self, word_list):
         """
         Got to have atleast one node in the tree - just stores one empty node
         """
         self.root = TrieNode("")
-        self.wordList = wordList
-        
-        with open(self.wordList) as f:
+        self.word_list = word_list
+        self.output = []
+
+        with open(self.word_list, encoding='utf-8') as f:
             for line in f:
                 self.insert(line)
-    
+
     def insert(self, word):
         """Insert a word into the trie"""
         node = self.root
-        
-        # Loops every character in the word and check if there is no child containing the char => creates a node at the current
+        # Loops every character in the word and check if there is no child containing the char =>
+        # creates a node at the current
         for char in word:
             if char in node.children:
                 node = node.children[char]
@@ -43,38 +51,37 @@ class SpellChecker(object):
                 new_node = TrieNode(char)
                 node.children[char] = new_node
                 node = new_node
-        
+
         # Mark the end of a word
         node.is_end = True
-
         # Increment the counter to indicate that we see this word once more
         node.counter += 1
-        
-    def cleanString(self, string):
+
+    def clean_string(self, string):
+        """ Removes strings that is exactly empty, a dash or a dot """
         # Resulting strings
         res = []
-        
         # Dont accept "-"
         for w in string.split('-'):
             if w not in res and w != '\n':
                 res.append(w)
-                
+
         # Dont accept whitespaces
         for w in string.split(' '):
             if w not in res and w != '\n':
                 res.append(w)
-        
+
         # Dont accept dots
         for w in string.split('.'):
             if w not in res and w != '\n':
                 res.append(w)
-        
+
         # Return list with valid cleaned strings
         return res
-        
+
     def dfs(self, node, prefix):
         """Depth-first traversal of the trie
-        
+
         Arguments:
             - node: the node to start with
             - prefix: the current prefix, for tracing a
@@ -82,18 +89,19 @@ class SpellChecker(object):
         """
         if node.is_end:
             self.output.append((prefix + node.char, node.counter))
-        
+
         for child in node.children.values():
             self.dfs(child, prefix + node.char)
-        
+
     def query(self, x):
         """
-        Given a prefix, get all words stored in the the trie with that prefix, sort them by the number of times they have been inserted
+        Given a prefix, get all words stored in the the trie with that prefix,
+        sort them by the number of times they have been inserted
         """
         # Keep all variants of the query - can be multiple different
         self.output = []
         node = self.root
-        
+
         # Check if the prefix is in the trie
         for char in x:
             if char in node.children:
@@ -101,30 +109,9 @@ class SpellChecker(object):
             else:
                 # cannot found the prefix, return empty list
                 return []
-        
+
         # Traverse the trie to get all candidates
         self.dfs(node, x[:-1])
 
         # Sort the results in reverse order and return
         return sorted(self.output, key=lambda x: x[1], reverse=True)
-    
-    
-    # def returnValidWords(self):
-    #     validWords = 0
-    #     invalidWords = 0
-    #     newLines = 0
-        
-    #     with open(self.outFile, "w") as invalidWordsFile:
-    #         with open(self.inputFile) as f:
-    #             for line in f:
-    #                 for word in self.cleanString(line):
-    #                     for splits in self.cleanString(word):
-    #                         if(splits == '\n'):
-    #                             newLines += 1
-    #                         elif(self.query(splits)):
-    #                             validWords += 1
-    #                         else:
-    #                             invalidWordsFile.write(splits)
-    #                             invalidWords += 1
-        
-    #     return f"Valid words: {validWords}\nInvalid words: {invalidWords}\nNew lines: {newLines}\n"
